2024 day 12 part 1 wip

+++ src/main/java/fr/ninauve/renaud/adventofcode/year2024/day12/Region.java
@@ -7,14 +7,14 @@ import java.util.stream.Stream;
 
 public record Region(CellContent content, Collection<Coordinates> coordinates) {
     public static Collection<Region> fromGrid(Grid<CellContent> grid) {
-        int previousId = -1;
-        final Map<Coordinates, Integer> regions = new HashMap<>();
-        final Map<Integer, Integer> aliases = new HashMap<>();
+        long previousId = -1;
+        final Map<Coordinates, Long> regions = new HashMap<>();
+        final Map<Long, Long> aliases = new HashMap<>();
         for (int row = 0; row < grid.getNbRows(); row++) {
             for (int col = 0; col < grid.getNbCols(); col++) {
                 Coordinates coordinates = new Coordinates(row, col);
                 CellContent cellContent = grid.get(coordinates);
-                Set<Integer> regionIds = Stream.of(coordinates.moveOf(Coordinates.UP), coordinates.moveOf(Coordinates.LEFT))
+                Set<Long> regionIds = Stream.of(coordinates.moveOf(Coordinates.UP), coordinates.moveOf(Coordinates.LEFT))
                         .filter(grid::isValid)
                         .filter(n -> cellContent.equals(grid.get(n)))
                         .map(regions::get)
@@ -24,15 +24,15 @@ public record Region(CellContent content, Collection<Coordinates> coordinates) {
                     previousId++;
                     regions.put(coordinates, previousId);
                 } else {
-                    List<Integer> sortedIds = regionIds.stream().sorted().toList();
-                    Integer regionId = sortedIds.getFirst();
+                    List<Long> sortedIds = regionIds.stream().sorted().toList();
+                    Long regionId = sortedIds.getFirst();
                     regions.put(coordinates, regionId);
                     sortedIds.stream().skip(1).forEach(alias -> aliases.put(alias, regionId));
                 }
             }
         }
 
-        Map<Integer, Set<Coordinates>> regionsById = regions.entrySet().stream()
+        Map<Long, Set<Coordinates>> regionsById = regions.entrySet().stream()
                 .map(e -> new AbstractMap.SimpleEntry<>(e.getKey(), regionId(aliases, e.getValue())))
                 .collect(Collectors.groupingBy(Map.Entry::getValue, Collectors.mapping(Map.Entry::getKey, Collectors.toSet())));
         return regionsById.values().stream()
@@ -40,9 +40,9 @@ public record Region(CellContent content, Collection<Coordinates> coordinates) {
                 .collect(Collectors.toSet());
     }
 
-    private static int regionId(Map<Integer, Integer> alias, int id) {
-        int previous = id;
-        Integer current = null;
+    private static long regionId(Map<Long, Long> alias, Long id) {
+        Long previous = id;
+        Long current = null;
         while((current = alias.get(previous)) != null) {
             previous = current;
         }
